Reorganized tests into packages
