+++ tests/src/cmput301w15t07/TravelTracker/test/model/InMemoryDataSourceTest.java
@@ -1,4 +1,4 @@
-package cmput301w15t07.TravelTracker.test;
+package cmput301w15t07.TravelTracker.test.model;
 
 import java.util.ArrayList;
 import java.util.Collection;
